# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AutoLlama v2.2 - Intelligent Context API
  description: |
    **AutoLlama v2.2 "Intelligent Context Llama"** - Enhanced RAG platform with intelligent contextual retrieval and advanced document understanding.
    
    ## 🚀 Enhanced Features v2.2
    - **35-60% Better Retrieval** through intelligent contextual processing
    - **Document Intelligence**: Automatic type detection and structural analysis
    - **Semantic Chunking**: Boundary-aware segmentation with hierarchical understanding
    - **Advanced Context Generation**: Document-aware prompts with retry logic
    - **Enhanced Metadata**: 11 new contextual fields for performance analysis
    - **Production-Ready**: Ultra-safe processing with comprehensive error handling
    
    ## Processing Capabilities
    - **Multi-format Processing**: PDF, DOCX, EPUB, CSV, URL fetching
    - **Hybrid Search**: Vector similarity + BM25 text search  
    - **Real-time Processing**: SSE streaming with progress updates
    - **Intelligent Chunking**: Semantic, structural, and hierarchical strategies
    - **Contextual Analysis**: Sentiment, topics, emotions, categories + document structure
    
    ## Authentication
    This API is designed for secure homelab deployment behind Tailscale VPN.
    
    ## Rate Limiting & Error Handling
    - **Search endpoints**: 100 requests/minute
    - **Processing endpoints**: 10 requests/5 minutes  
    - **Health endpoints**: No limits
...
  version: 2.2.0
...
  contact:
...
servers:
...
paths:
...
  /search:
    get:
      tags: [Search & Retrieval]
      summary: Hybrid search
      description: |
        Perform hybrid search combining vector similarity and BM25 text search.
        Returns ranked results with similarity scores and contextual information.
      parameters:
        - name: q
          in: query
          required: true
          description: "🔍 Search query text - Try: 'machine learning', 'python tutorial', or any topic"
          schema:
            type: string
            minLength: 1
            maxLength: 1000
          example: "machine learning tutorial"
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of results to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: includeChunks
          in: query
          description: Include related chunks in results
          schema:
            type: boolean
            default: false
        - name: threshold
          in: query
          description: Minimum similarity threshold (0.0-1.0)
          schema:
            type: number
            minimum: 0.0
            maximum: 1.0
            default: 0.7
      responses:
        '200':
          description: Search results with similarity scores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /search/documents:
    get:
      tags: [Search & Retrieval]
      summary: Document search
      description: Search documents by title, content, and metadata
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [created_at, updated_at, title, relevance]
            default: relevance
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Document search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSearchResponse'
...
  /documents:
    get:
      tags: [Document Management]
      summary: List documents
      description: Get paginated list of documents with optional filtering
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          description: Search term for filtering
          schema:
            type: string
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [created_at, updated_at, title, chunk_count]
            default: created_at
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Paginated document list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
  /documents/{id}:
    get:
      tags: [Document Management]
      summary: Get document details
      description: Retrieve specific document with chunks and metadata
      parameters:
        - name: id
          in: path
          required: true
          description: Document ID
          schema:
            type: string
            format: uuid
        - name: chunkLimit
          in: query
          description: Maximum chunks to include
          schema:
            type: integer
            minimum: 0
            maximum: 100
            default: 50
        - name: chunkOffset
          in: query
          description: Chunk offset for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Document details with chunks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetailsResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Document Management]
      summary: Update document
      description: Update document metadata and properties
      parameters:
        - name: id
          in: path
          required: true
          description: Document ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentUpdateRequest'
      responses:
        '200':
          description: Document updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUpdateResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Document Management]
      summary: Delete document
      description: |
        Delete document and all associated chunks and vector data.
        This operation is irreversible and performed atomically.
      parameters:
        - name: id
          in: path
          required: true
          description: Document ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDeleteResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{id}/chunks:
    get:
      tags: [Document Management]
      summary: Get document chunks
      description: Retrieve paginated chunks for a specific document
      parameters:
        - name: id
          in: path
          required: true
          description: Document ID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: includeEmbeddings
          in: query
          description: Include embedding vectors in response
          schema:
            type: boolean
            default: false
        - name: search
          in: query
          description: Search within chunks
          schema:
            type: string
      responses:
        '200':
          description: Document chunks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentChunksResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "⚠️ NOT IMPLEMENTED - Leave empty for local development"
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "⚠️ NOT IMPLEMENTED - Leave empty for local development"
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: "Validation failed"
              code: "VALIDATION_ERROR"
              details:
                - field: "chunkSize"
                  message: "Must be between 100 and 5000"
            statusCode: 400
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: "Document not found"
              code: "NOT_FOUND"
              timestamp: "2024-01-15T10:30:00.000Z"
            statusCode: 404
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: "Internal server error"
              code: "INTERNAL_ERROR"
              timestamp: "2024-01-15T10:30:00.000Z"
            statusCode: 500
...
  schemas:
...
    SearchResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        query:
          type: string
          example: "artificial intelligence"
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        total:
          type: integer
          description: Total number of results
        limit:
          type: integer
        offset:
          type: integer
        searchType:
          type: string
          enum: [hybrid, vector, text]
          example: "hybrid"
        processingTime:
          type: number
          description: Search processing time in milliseconds
    SearchResult:
      type: object
      properties:
        chunk_id:
          type: string
          format: uuid
        chunk_text:
          type: string
        title:
          type: string
        url:
          type: string
          format: uri
        chunk_index:
          type: integer
        similarity_score:
          type: number
          minimum: 0
          maximum: 1
        sentiment:
          type: string
          enum: [positive, negative, neutral]
        contextual_summary:
          type: string
        main_topics:
          type: array
          items:
            type: string
        category:
          type: string
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/ChunkData'
          description: Related chunks (if includeChunks=true)
...
    DocumentSearchResponse:
      type: object
      properties:
        success:
          type: boolean
        query:
          type: string
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSummary'
        total:
          type: integer
        searchType:
          type: string
          example: "document"
...
    DocumentListResponse:
      type: object
      properties:
        success:
          type: boolean
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
    DocumentSummary:
      type: object
      properties:
        document_id:
          type: string
          format: uuid
        title:
          type: string
        url:
          type: string
          format: uri
          nullable: true
        chunk_count:
          type: integer
        status:
          type: string
          enum: [processing, completed, failed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
    DocumentDetailsResponse:
      type: object
      properties:
        success:
          type: boolean
        document:
          $ref: '#/components/schemas/Document'
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/ChunkData'
        chunksPagination:
          type: object
          properties:
            total:
              type: integer
            returned:
              type: integer
    Document:
      type: object
      properties:
        document_id:
          type: string
          format: uuid
        title:
          type: string
        url:
          type: string
          format: uri
          nullable: true
        filename:
          type: string
          nullable: true
        chunk_count:
          type: integer
        status:
          type: string
          enum: [processing, completed, failed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          properties:
            contentType:
              type: string
            wordCount:
              type: integer
            language:
              type: string
            fileSize:
              type: integer
            processingTime:
              type: number
    DocumentUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        metadata:
          type: object
          properties:
            category:
              type: string
            tags:
              type: array
              items:
                type: string
            priority:
              type: string
              enum: [low, medium, high]
            notes:
              type: string
    DocumentUpdateResponse:
      type: object
      properties:
        success:
          type: boolean
        document:
          $ref: '#/components/schemas/Document'
        message:
          type: string
    DocumentDeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        deletedChunks:
          type: integer
        vectorEntriesDeleted:
          type: integer
    DocumentChunksResponse:
      type: object
      properties:
        success:
          type: boolean
        documentId:
          type: string
          format: uuid
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/ChunkData'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    ChunkData:
      type: object
      properties:
        chunk_id:
          type: string
          format: uuid
        chunk_text:
          type: string
        chunk_index:
          type: integer
        sentiment:
          type: string
          enum: [positive, negative, neutral]
        emotions:
          type: array
          items:
            type: string
        category:
          type: string
        main_topics:
          type: array
          items:
            type: string
        contextual_summary:
          type: string
        similarity_score:
          type: number
          minimum: 0
          maximum: 1
        embedding:
          type: array
          items:
            type: number
          description: Embedding vector (only if requested)

    # Processing Schemas
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              nullable: true
            details:
              oneOf:
                - type: string
                - type: object
                - type: array
              nullable: true
        statusCode:
          type: integer

# No global security required - this is a local development API